feat(jinja): deprecate contrib module on v2
