+++ docs/reference/contrib/jinja.rst
@@ -1,5 +1,8 @@
 jinja
 =====
 
+.. deprecated:: 2.22.0
+    Import from :mod:`litestar.plugins.jinja` instead. This module will be removed in 3.0.0.
+
 .. automodule:: litestar.contrib.jinja
     :members:
