+++ docs/reference/plugins/jinja.rst
@@ -1,0 +1,6 @@
+=====
+jinja
+=====
+
+.. automodule:: litestar.plugins.jinja
+    :members:
